server/pyproject.toml
[project]
name = "annotated-paper"
version = "0.1.0"
requires-python = ">= 3.12"
description = "An Annotated Paper Application"
readme = "README.md"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic[email]>=2.10",
    "python-dotenv>=1.0.1",
    "google-genai>=1.7.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.39",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.15.1",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.8.1",
    "types-psycopg2>=2.9.21.20250318",
    "types-requests>=2.32.0.20250306",
    "markitdown[pdf]>=0.1.0",
    "sqlalchemy2-stubs>=0.0.2a38",
]
